Rename conlink to conlink.py (and conlink.py to *_inner.py)
